flask-server/Ai/requirements.txt
fastapi
uvicorn
torch
transformers
numpy
python-dotenv